Correct c_uint/c_uint32↔c_size_t prototype mismatches
In SIDX_NewBuffer and Index_InsertTPData bindings, fix use of c_uint
(unsigned int) or c_uint32 (uint32_t) for value parameters that should be
c_size_t (size_t) for consistency with the libspatialite C API.

https://github.com/Toblerity/rtree/issues/220#issuecomment-1046256534

+++ rtree/core.py
@@ -527,7 +527,7 @@ rt.IndexProperty_GetIndexID.argtypes = [ctypes.c_void_p]
 rt.IndexProperty_GetIndexID.restype = ctypes.c_int64
 rt.IndexProperty_GetIndexID.errcheck = check_value  # type: ignore
 
-rt.SIDX_NewBuffer.argtypes = [ctypes.c_uint]
+rt.SIDX_NewBuffer.argtypes = [ctypes.c_size_t]
 rt.SIDX_NewBuffer.restype = ctypes.c_void_p
 rt.SIDX_NewBuffer.errcheck = check_void  # type: ignore
 
@@ -551,7 +551,7 @@ try:
         ctypes.c_double,
         ctypes.c_uint32,
         ctypes.POINTER(ctypes.c_ubyte),
-        ctypes.c_uint32,
+        ctypes.c_size_t,
     ]
     rt.Index_InsertTPData.restype = ctypes.c_int
     rt.Index_InsertTPData.errcheck = check_return  # type: ignore
